requirements.txt
flask
flask_cors
SQLAlchemy
python-dotenv
PyMySQL